feat: Add config of security
